Change signup configuration read when retrieving consent purposes
