+++ features/apimgt/org.wso2.carbon.apimgt.store.feature/src/main/resources/store/site/blocks/user/sign-up/block.jag
@@ -5,9 +5,9 @@ jagg.block("user/sign-up", {
     },
 
 	getOutputs:function () {
-        var APIUtil = org.wso2.carbon.apimgt.impl.utils.APIUtil;
+        var SelfSignUpUtil = org.wso2.carbon.apimgt.impl.utils.SelfSignUpUtil;
         var request_tenant = jagg.getTenantDomain();
-        var purposes = APIUtil.getConsentPurposes(request_tenant);
+        var purposes = SelfSignUpUtil.getConsentPurposes(request_tenant);
         var result, fields, user = jagg.module("user");
         result = user.getUserFields();
         fields = result.fields;
